requirements.txt
python-dotenv
Flask
gunicorn
spotipy
pandas
numpy
gspread
oauth2client
